backend/requirements.txt
# Backend Python Dependencies
# For AI-Awareness Project

# Django Framework
Django==5.2.9
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1

# Two-Factor Authentication
django-otp==1.5.4
qrcode==7.4.2
pyotp==2.9.0
cryptography==42.0.0

# CORS Support
django-cors-headers==4.9.0

# HTTP Requests
requests==2.32.5

# Database
sqlparse==0.5.4

# PostgreSQL Driver
psycopg2-binary==2.9.9

# Security & Utilities
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.0

# Optional: For production deployment
gunicorn==23.0.0

# Optional: For better logging
python-dotenv==1.0.1


